Added user profiles to the registration page/header/profile page. styled the profile page and resized all the images to fit.
