+++ src/components/Header/header.css
@@ -1,0 +1,9 @@
+.signout-button{
+    height: 36.5px;
+}
+
+.head-img-profile{
+    height: 36.5px;
+    width: 36.5px;
+    border-radius: 50%;
+}
