+++ src/components/profile/passwordform.css
@@ -1,0 +1,12 @@
+.password-form{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+.password-input{
+background-color: white;
+border-radius: 20px;
+width: 500px;
+}
