+++ src/components/profile/profile.css
@@ -17,3 +17,24 @@
 .add-user-to-cohort-btn:hover {
   background-color: #1862ac;
 }
+
+.profile-form{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 5px;
+}
+
+.profile-user-text{
+  background-color: white;
+  border-radius: 20px;
+ 
+}
+
+.img-profile{
+  margin: 40px;
+  border-radius: 10px;
+  height: 200px;
+  width: 200px;
+}
