+++ src/components/searchBar/RenderListOfStudents.js
@@ -13,13 +13,17 @@ const RenderListOfStudents = () => {
       </header>
       <ul className='students-list'>
         {userDataToRender.map((student, index) => {
-          const { first_name, last_name } = student;
+          const { first_name, last_name, profile_url } = student;
           return (
             <li className='student-list-item' key={index}>
               <Link to={`/profile/${student.id}`} className='link'>
+                <div className='two-column-grid'>
+                <img className='list-of-students-img' src={profile_url}/>
                 <p>
-                  {first_name} {last_name}
+                 {first_name} {last_name}
                 </p>
+
+                </div>
               </Link>
             </li>
           );
